+++ Src/EonSharp/Api/Bot/BotEndpoint.cs
@@ -95,9 +95,9 @@ namespace EonSharp.Api.Bot
 			await m_transportClient.ProcessCommandAsync(EndpointUrl, new RpcRequest("transactions.putTransaction", new object[] { tx }, Interlocked.Increment(ref m_id)));
 		}
 
-		async Task<ColoredCoinInfo> IColoredCoin.GetInfo(string id, int timestamp)
+		async Task<ColoredCoinInfo> IColoredCoin.GetInfo(string id)
 		{
-			var res = await m_transportClient.ProcessCommandAsync(EndpointUrl, new RpcRequest("coloredCoin.getInfo", new object[] { id, timestamp }, Interlocked.Increment(ref m_id)));
+			var res = await m_transportClient.ProcessCommandAsync(EndpointUrl, new RpcRequest("coloredCoin.getInfo", new object[] { id }, Interlocked.Increment(ref m_id)));
 			return (res.Result as JObject)?.ToObject<ColoredCoinInfo>();
 		}
 	}